+++ mitra_models/migrations/V0079__activitypub_object.sql
@@ -1,0 +1,5 @@
+CREATE TABLE activitypub_object (
+    object_id VARCHAR(2000) PRIMARY KEY,
+    object_data JSONB NOT NULL,
+    created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP
+);