+++ src/job_queue/periodic_tasks.rs
@@ -7,6 +7,7 @@ use mitra_activitypub::queues::{
 use mitra_adapters::media::delete_orphaned_media;
 use mitra_config::Config;
 use mitra_models::{
+    activitypub::queries::delete_activitypub_objects,
     background_jobs::queries::{
         delete_job_from_queue,
         get_job_batch,
@@ -170,6 +171,21 @@ pub async fn prune_remote_emojis(
         let deletion_queue = delete_emoji(db_client, &emoji_id).await?;
         delete_orphaned_media(config, db_client, deletion_queue).await?;
         log::info!("deleted unused emoji {}", emoji_id);
+    };
+    Ok(())
+}
+
+pub async fn prune_activitypub_objects(
+    _config: &Config,
+    db_pool: &DatabaseConnectionPool,
+) -> Result<(), Error> {
+    const CACHE_EXPIRATION_DAYS: u32 = 5;
+    let db_client = &**get_database_client(db_pool).await?;
+    let created_before = days_before_now(CACHE_EXPIRATION_DAYS);
+    let deleted_count =
+        delete_activitypub_objects(db_client, created_before).await?;
+    if deleted_count > 0 {
+        log::info!("deleted {deleted_count} activitypub objects");
     };
     Ok(())
 }